+++ src/core/bezier.ts
